requirements.txt
typer~=0.12.3
pathlib~=1.0.1
nxt-python~=3.3.0
rich~=13.7.1
setuptools~=69.5.1